增加回收站和文件管理功能，删除主栏目提示有子栏目先删除子栏目，删除子栏目提示该子栏目有文章先删除文章；

+++ Application/Admin/Controller/CategoryController.class.php
@@ -11,10 +11,13 @@ class CategoryController extends CommonController
     {
         parent::__construct();
         $this->category = D("Category");
+        $this->article = D("Article");
     }
 
     public function index()
     {
+
+
         $categories = $this->category->all();
         $this->assign("categories", $categories);
         $this->display();
@@ -22,7 +25,8 @@ class CategoryController extends CommonController
 
     public function create()
     {
-        $categories=$this->category->all();
+
+        $categories = $this->category->all();
         $this->assign("categories", $categories);
         $this->display();
 
@@ -30,7 +34,16 @@ class CategoryController extends CommonController
 
     public function destroy()
     {
-        $id =I("post.id");
+        $id = I("post.id");
+
+        $articles = $this->article->where("category_id='$id'")->select();
+        if ($articles){
+            $this->success('必须先删除该栏目的文章才可以删除该栏目');
+            return false;
+        }
+//        dump($articles);
+//        exit;
+        //$this->assign('articles', $articles);
         $this->category->delete($id);
         $this->success('删除成功');
     }
@@ -38,9 +51,8 @@ class CategoryController extends CommonController
     public function sort_order()
     {
         $id = I("post.id");
-        $sort_order =I("post.sort_order");
-        foreach ($id as $k=>$v)
-        {
+        $sort_order = I("post.sort_order");
+        foreach ($id as $k => $v) {
             $this->category->where("id='$v'")->setField("sort_order", $sort_order[$k]);
         }
         $this->redirect('index');
@@ -57,23 +69,25 @@ class CategoryController extends CommonController
         $this->redirect('index');
 
     }
+
     public function edit()
     {
-        $id=I("get.id");
+        $id = I("get.id");
 //        dump($id);
 //        exit();
-        $categories=$this->category->all();
+        $categories = $this->category->all();
         $this->assign("categories", $categories);
-        $category=$this->category->where("id='$id'")->find();
+        $category = $this->category->where("id='$id'")->find();
         $this->assign('category', $category);
 
         $this->display();
     }
+
     public function update()
     {
 //        dump($_POST);
 //        exit();
-        $Category=M("Category");
+        $Category = M("Category");
         $Category->create();
         $Category->save();
         F('categories', NULL);
